Stage 4
Completed basic outline of stage 4. Will need to add signs, and background. Need signifier for the wind change as it is very infuriating to not be able to counter it with good aiming. I think the aiming should be an arc instead of a straight line as it can be hard for people to understand how heavy the ball is, and thus how much the ball will arc (because of gravity).